refactor: 파일 구조 변경
- Car 클래스를 domain 패키지로 이동
- Game 클래스를 service 패키지로 이동
- InputUtils, OutputUtils, Constant 클래스를 util 패키지로 이동
